guilib/GUIInfoTypes: move each class to a dedicated file
Reduce include dependencies.
