+++ xbmc/guilib/GUIControl.h
@@ -32,7 +32,8 @@
 #include "utils/Color.h"
 #include "windowing/GraphicContext.h" // needed by any rendering operation (all controls)
 #include "VisibleEffect.h"  // needed for the CAnimation members
-#include "guiinfo/GUIInfoTypes.h" // needed for CGUIInfoColor to handle infolabel'ed colors
+#include "guiinfo/GUIInfoColor.h" // needed for CGUIInfoColor to handle infolabel'ed colors
+#include "guiinfo/GUIInfoBool.h"
 #include "DirtyRegion.h"
 
 class CGUIListItem; // forward
