+++ xbmc/guilib/GUIFadeLabelControl.h
@@ -32,6 +32,7 @@
 
 #include "GUIControl.h"
 #include "GUILabel.h"
+#include "guilib/guiinfo/GUIInfoLabel.h"
 
 /*!
  \ingroup controls
