+++ xbmc/guilib/GUIIncludes.cpp
@@ -21,7 +21,7 @@
 #include "GUIIncludes.h"
 #include "addons/Skin.h"
 #include "GUIInfoManager.h"
-#include "guiinfo/GUIInfoTypes.h"
+#include "guilib/guiinfo/GUIInfoLabel.h"
 #include "guilib/GUIComponent.h"
 #include "utils/log.h"
 #include "utils/XBMCTinyXML.h"
