+++ xbmc/guilib/GUILabelControl.h
@@ -30,6 +30,7 @@
 
 #include "GUIControl.h"
 #include "GUILabel.h"
+#include "guilib/guiinfo/GUIInfoLabel.h"
 
 /*!
  \ingroup controls
