+++ xbmc/guilib/GUIListItemLayout.h
@@ -22,7 +22,7 @@
 
 #include "GUIListGroup.h"
 #include "GUITexture.h"
-#include "guiinfo/GUIInfoTypes.h"
+#include "guilib/guiinfo/GUIInfoLabel.h"
 
 class CGUIListItem;
 class CFileItem;
