+++ xbmc/guilib/GUIListLabel.h
@@ -27,6 +27,7 @@
 
 #include "GUIControl.h"
 #include "GUILabel.h"
+#include "guilib/guiinfo/GUIInfoLabel.h"
 
 /*!
  \ingroup controls
